@@ -1,7 +1,5 @@
 #pragma once
 
-#define UUID_SYSTEM_GENERATOR
-
 #include <chrono>
 #include <filesystem>
 #include <memory>
@@ -16,6 +14,8 @@
 #include <fstream>
 #include <array>
 #include <vector>
+#include <list>
+#include <deque>
 #include <unordered_map>
 #include <unordered_set>
 #include <regex>
@@ -26,9 +26,18 @@
 #include <nameof.hpp>
 #include <stduuid/uuid.h>
 #include <fmt/format.h>
+#include <fmt/chrono.h>
+#include <nlohmann/json.hpp>
 
 #include "Macros.h"
 #include "Reflect.h"
+
+#include "ChronoUtils.h"
+#include "StringUtils.h"
+#include "UUIDUtils.h"
+#include "FormatUtils.h"
+#include "PathUtils.h"
+#include "ValidationUtils.h"
 
 namespace Calyx {
 
@@ -36,11 +45,24 @@ namespace Calyx {
 
     using String = std::string;
 
+    using StringView = std::string_view;
+
+    using StringStream = std::stringstream;
+
     template<typename T>
     using List = std::vector<T>;
 
+    template<typename T>
+    using LinkedList = std::list<T>;
+
+    template<typename T>
+    using Deque = std::deque<T>;
+
     template<typename K, typename V>
     using Map = std::unordered_map<K, V>;
+
+    template<typename K, typename V>
+    using OrderedMap = std::map<K, V>;
 
     template<typename T>
     using Set = std::unordered_set<T>;
@@ -58,12 +80,15 @@ namespace Calyx {
     using WeakRef = std::weak_ptr<T>;
 
     template<typename T>
+    using Optional = std::optional<T>;
+
+    template<typename T>
     using Function = std::function<T>;
 
     namespace FileSystem = std::filesystem;
     using Path = std::filesystem::path;
 
-    using IdentifierT = entt::id_type;
+    using IdentifierType = entt::id_type;
 
     namespace ThisThread = std::this_thread;
     using Thread = std::thread;
@@ -73,11 +98,13 @@ namespace Calyx {
         using uint16 = uint16_t;
         using uint32 = uint32_t;
         using uint64 = uint64_t;
+        typedef unsigned __int128 uint128;
 
         using int8 = int8_t;
         using int16 = int16_t;
         using int32 = int32_t;
         using int64 = int64_t;
+        typedef __int128 int128;
     }
 
     namespace Math {
@@ -118,18 +145,14 @@ namespace Calyx {
 
     namespace FileSystem = std::filesystem;
 
-    namespace Reflect {
-        using TypeID = entt::id_type;
+    template<typename T, typename ... Args>
+    inline constexpr Scope<T> CreateScope(Args&& ... args) {
+        return std::make_unique<T>(std::forward<Args>(args)...);
     }
 
     template<typename T, typename ... Args>
-    constexpr Scope<T> CreateScope(Args&& ... args) {
-        return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
-    }
-
-    template<typename T, typename ... Args>
-    constexpr Ref<T> CreateRef(Args&& ... args) {
-        return std::shared_ptr<T>(new T(std::forward<Args>(args)...));
+    inline constexpr Ref<T> CreateRef(Args&& ... args) {
+        return std::make_shared<T>(std::forward<Args>(args)...);
     }
 
 }
@@ -139,3 +162,5 @@ using namespace Calyx::Integer;
 using namespace Calyx::Math;
 using namespace Calyx::Chrono;
 using namespace Calyx::Reflect;
+
+using nlohmann::json;
