@@ -2,17 +2,23 @@
 
 #include "assets/Asset.h"
 
+namespace Calyx {
+    class GameObject;
+}
+
 namespace Calyx::Editor {
 
     class InspectorGUI {
 
     public:
-        static bool GameAssetControl(const String& name, AssetT assetType, Ref<IAsset>& value);
+        static bool ComponentPicker(GameObject* gameObject, const Set<AssetType>& excluded);
 
-        static bool TextControl(const String& name, String& value);
+        static bool GameAssetControl(const String& name, AssetType assetType, Ref<IAsset>& value);
+
+        static bool TextControl(const String& name, String& value, bool readonly = false);
         static bool Vec3Control(const String& name, vec3& value, float speed = 1.0f);
         static bool Vec2Control(const String& name, vec2& value, float speed = 1.0f);
-        static bool SliderControl(const String& name, float& value, float min = 0.0f, float max = 0.0f);
+        static bool FloatControl(const String& name, float& value, float speed = 1.0f, float min = 0.0f, float max = 0.0f);
 
         static bool BeginPropertyTable(const String& name);
         static void EndPropertyTable();
@@ -20,10 +26,6 @@ namespace Calyx::Editor {
         static void Property(const String& label);
 
     private:
-        static String s_assetSearch;
-        static List<AssetRegistry::AssetMeta> s_assetList;
-        static int s_assetSelected;
-
         static bool DragFloatN_Colored(
             const char* label,
             float* v,
