@@ -8,8 +8,8 @@ namespace Calyx::Editor {
 
     class TransformComponentInspector : public TypeInspector<TransformComponent> {
         CX_BEGIN_REFLECT();
-        void CX_ON_INSPECTOR_CONTEXT_GUI(IdentifierT typeID, TransformComponent& tc);
-        void CX_ON_INSPECTOR_GUI(IdentifierT typeID, TransformComponent& tc);
+        void CX_ON_INSPECTOR_CONTEXT_GUI(TransformComponent& tc);
+        void CX_ON_INSPECTOR_GUI(TransformComponent& tc);
     };
 
 }
