@@ -2,7 +2,7 @@
 
 namespace Calyx::Editor {
 
-    void StringInspector::CX_ON_INSPECTOR_GUI(IdentifierT typeID, String& value) {
+    void StringInspector::CX_ON_INSPECTOR_GUI(String& value) {
         InspectorGUI::TextControl("##text", value);
     }
 
