@@ -2,7 +2,7 @@
 
 namespace Calyx::Editor {
 
-    void Vec2Inspector::CX_ON_INSPECTOR_GUI(IdentifierT typeID, vec2& value) {
+    void Vec2Inspector::CX_ON_INSPECTOR_GUI(vec2& value) {
         InspectorGUI::Vec2Control("##vec2", value);
     }
 
