@@ -2,7 +2,7 @@
 
 namespace Calyx::Editor {
 
-    void Vec3Inspector::CX_ON_INSPECTOR_GUI(IdentifierT typeID, vec3& value) {
+    void Vec3Inspector::CX_ON_INSPECTOR_GUI(vec3& value) {
         InspectorGUI::Vec3Control("##vec3", value);
     }
 
