@@ -1,62 +1,68 @@
 #pragma once
 
-#include "ecs/Component.h"
-#include "ecs/components/TransformComponent.h"
 #include "scene/Scene.h"
 
 namespace Calyx {
 
-    class GameObject {
+    class IComponent;
+    class SceneExporter;
+    class SceneImporter;
+
+    class CALYX_API GameObject {
 
         friend class Scene;
+        friend class SceneExporter;
+        friend class SceneImporter;
         friend class Editor::EditorLayer;
 
         template<typename T, typename ... Args>
         friend constexpr Scope<T> Calyx::CreateScope(Args&& ... args);
 
     private:
-        GameObject(Scene* scene, entt::entity entityID, const String& name);
+        GameObject(Scene* scene, UUID id, entt::entity entityID, const String& name);
 
     public:
-        template<typename T, typename ...Ts>
-        T* AddComponent(Ts ...args) {
-            T& component = m_scene->m_entityRegistry.emplace<T>(m_entityID, std::forward<Ts>(args)...);
+        template<typename T, typename ...Args>
+        T* AddComponent(Args&&... args) {
+            static_assert(std::is_base_of_v<IComponent, T>, "T must extend IComponent!");
+            T& component = m_scene->m_entityRegistry.emplace<T>(m_entityID, std::forward<Args>(args)...);
             component.m_gameObject = this;
             return &component;
         }
 
-        void AddComponent(const entt::meta_type& type) {
-            auto storage = m_scene->m_entityRegistry.storage(type.info().hash());
-            storage->emplace(m_entityID);
-            void* component = storage->get(m_entityID);
-            auto ref = type.from_void(component);
-            ref.assign(type.construct());
-        }
+        void AddComponent(const entt::meta_type& type);
 
         template<typename T>
         void RemoveComponent() {
             m_scene->m_entityRegistry.erase<T>(m_entityID);
         }
 
+        void RemoveComponent(const entt::meta_type& type);
+
         template<typename T>
         T* GetComponent() {
             if (HasComponent<T>())
-                return &m_scene->m_entityRegistry.template get<T>(m_entityID);
+                return &m_scene->m_entityRegistry.get<T>(m_entityID);
             return nullptr;
         }
 
+        bool HasComponent(const entt::meta_type& type);
+
         template<typename T>
         bool HasComponent() {
-            return m_scene->m_entityRegistry.template all_of<T>(m_entityID);
+            return m_scene->m_entityRegistry.all_of<T>(m_entityID);
         }
 
         void AddChild(GameObject* child);
         void SetParent(GameObject* parent);
         void DetachFromParent();
 
+        void SetName(const String& name) { m_name = name; }
+
+        const UUID& GetID() const { return m_id; }
+
         const Scene* GetScene() const { return m_scene; }
-        String GetName() const { return m_name; }
-        String& GetNameRef() { return m_name; }
+        const String& GetName() const { return m_name; }
         const Set<GameObject*>& GetChildren() const { return m_children; }
         GameObject* GetParent() const { return m_parent; }
 
@@ -67,6 +73,7 @@ namespace Calyx {
 
     private:
         Scene* m_scene;
+        UUID m_id;
         entt::entity m_entityID = entt::null;
         String m_name;
         GameObject* m_parent = nullptr;
