@@ -7,23 +7,25 @@
 #include <ImGuizmo.h>
 #include <implot.h>
 
-#include "core/ILayer.h"
+#include "core/Layer.h"
 #include "core/Application.h"
+#include "ui/Definitions.h"
 
 namespace Calyx {
 
-    class GuiLayer : public ILayer {
+    class CALYX_API GuiLayer : public ILayer {
 
     public:
-        virtual void OnAttach() override;
-        virtual void OnDetach() override;
-        virtual void OnEvent(Event& event) override;
+        void OnAttach() override;
+        void OnDetach() override;
+        void OnEvent(Event& event) override;
 
         void Begin();
         void End();
 
-        virtual String GetName() const override { return "GUI Layer"; };
+        String GetName() const override { return "GUI Layer"; };
 
+    private:
         static void SetupImGuiTheme();
 
     private:
