@@ -1,8 +1,10 @@
 #include "assets/AssetRegistry.h"
 
+#include "ecs/Component.h"
 #include "render/objects/Shader.h"
 #include "render/objects/Texture2D.h"
 #include "assets/Mesh.h"
+#include "reflect/ClassRegistry.h"
 
 namespace Calyx {
 
@@ -11,7 +13,7 @@ namespace Calyx {
     AssetRegistry::AssetRegistry()
         : m_assetWatcher(CreateScope<efsw::FileWatcher>()) {
         InitAssetTypes();
-        _AddSearchPath("assets");
+        RegisterCoreComponents();
         m_assetWatcher->watch();
     }
 
@@ -48,7 +50,7 @@ namespace Calyx {
 
         // Load asset by type
         CX_MAP_FIND(m_assetMeta, id, i_assetMeta) {
-            CX_MAP_FIND(m_assetTypes, i_assetMeta->second.type, i_assetType) {
+            CX_MAP_FIND(m_registeredAssetTypes, i_assetMeta->second.type, i_assetType) {
                 auto* asset = i_assetType->second.loadAssetFn(i_assetMeta->second.path.string());
                 if (asset == nullptr)
                     return Ref<IAsset>();
@@ -97,9 +99,16 @@ namespace Calyx {
         }
     }
 
-    void AssetRegistry::_SearchAssets(AssetT assetType, const String& query, List<AssetMeta>& outList) {
+    void AssetRegistry::_SearchComponents(const String& query, const Set<AssetType>& excluded, List<AssetMeta>& outList) {
         for (const auto& [_, meta]: m_assetMeta) {
-            if (meta.type == assetType && Utils::IsSearchMatch(query, meta.displayName))
+            if (meta.isComponent && !excluded.contains(meta.type) && StringUtils::IsSearchMatch(query, meta.displayName))
+                outList.push_back(meta);
+        }
+    }
+
+    void AssetRegistry::_SearchAssets(AssetType assetType, const String& query, List<AssetMeta>& outList) {
+        for (const auto& [_, meta]: m_assetMeta) {
+            if (meta.type == assetType && StringUtils::IsSearchMatch(query, meta.displayName))
                 outList.push_back(meta);
         }
     }
@@ -116,29 +125,30 @@ namespace Calyx {
     }
 
     String AssetRegistry::_GetAssetDisplayName(const IAsset* ptr) {
-        if (ptr == nullptr) return "None";
-        CX_MAP_FIND(m_assetPointers_IDs, ptr, i_assetID) {
-            CX_MAP_FIND(m_assetMeta, i_assetID->second, i_assetMeta) {
-                return i_assetMeta->second.displayName;
-            }
-        }
-        return "";
+        if (ptr == nullptr)
+            return "None";
+
+        auto meta = _GetAssetMeta(ptr);
+        if (meta == nullptr)
+             return "";
+        
+        return meta->displayName;
     }
 
     String AssetRegistry::_GetAssetDisplayName(const Ref<IAsset>& ref) {
         return _GetAssetDisplayName(ref.get());
     }
 
-    AssetRegistry::AssetMeta AssetRegistry::_GetAssetMeta(const IAsset* ptr) {
+    const AssetRegistry::AssetMeta* AssetRegistry::_GetAssetMeta(const IAsset* ptr) {
         CX_MAP_FIND(m_assetPointers_IDs, ptr, i_assetID) {
             CX_MAP_FIND(m_assetMeta, i_assetID->second, i_assetMeta) {
-                return i_assetMeta->second;
+                return &i_assetMeta->second;
             }
         }
-        return {};
+        return nullptr;
     }
 
-    AssetRegistry::AssetMeta AssetRegistry::_GetAssetMeta(const Ref<IAsset>& ref) {
+    const AssetRegistry::AssetMeta* AssetRegistry::_GetAssetMeta(const Ref<IAsset>& ref) {
         return _GetAssetMeta(ref.get());
     }
 
@@ -146,7 +156,7 @@ namespace Calyx {
         return Path(path).lexically_normal().string();
     }
 
-    bool AssetRegistry::FindAssetType(const Path& file, AssetT& outType) {
+    bool AssetRegistry::FindAssetType(const Path& file, AssetType& outType) {
         String ext = file.extension().string();
         CX_MAP_FIND(m_assetExtensions_Types, ext, i_assetType) {
             outType = i_assetType->second;
@@ -155,8 +165,8 @@ namespace Calyx {
         return false;
     }
 
-    bool AssetRegistry::ExpectAssetType(const Path& file, AssetT expectedType) {
-        AssetT actualType;
+    bool AssetRegistry::ExpectAssetType(const Path& file, AssetType expectedType) {
+        AssetType actualType;
         if (!FindAssetType(file, actualType))
             return false;
         return expectedType == actualType;
@@ -202,20 +212,83 @@ namespace Calyx {
 
             auto relativeFilePath = FileSystem::relative(file, searchPath.path);
             auto assetName = NormalizePath(relativeFilePath.string());
-            AssetT assetType;
+            AssetType assetType;
             if (!FindAssetType(file, assetType))
                 continue;
 
-            UUID assetID = Utils::GenerateUUID();
-            m_assetNames_IDs[assetName] = assetID;
+            UUID assetID;
+            json meta;
+            if (LoadMetaFile(file.path(), meta)) {
+                String id = meta.at("id");
+                assetID = UUID::from_string(id).value_or(UUIDUtils::Generate());
+                m_assetNames_IDs[assetName] = assetID;
+                m_assetMeta[assetID] = {
+                    .id = assetID,
+                    .type = assetType,
+                    .name = assetName,
+                    .displayName = meta.at("display_name"),
+                    .path = FileSystem::absolute(file)
+                };
+            } else {
+                assetID = UUIDUtils::Generate();
+                m_assetNames_IDs[assetName] = assetID;
+                m_assetMeta[assetID] = {
+                    .id = assetID,
+                    .type = assetType,
+                    .name = assetName,
+                    .displayName = relativeFilePath.stem().string(),
+                    .path = FileSystem::absolute(file)
+                };
+                WriteMetaFile(file.path(), m_assetMeta[assetID]);
+            }
+
+            searchPath.assets.push_back(assetID);
+        }
+    }
+
+    Path AssetRegistry::GetMetaFile(const Path& assetFile) {
+        return assetFile.parent_path() / (assetFile.stem().string() + ".meta");
+    }
+
+    bool AssetRegistry::LoadMetaFile(const Path& assetFile, json& meta) {
+        Path metaFile = GetMetaFile(assetFile);
+        if (!FileSystem::exists(metaFile) || !FileSystem::is_regular_file(metaFile))
+            return false;
+        std::ifstream stream(metaFile);
+        meta = json::parse(stream);
+        return true;
+    }
+
+    void AssetRegistry::WriteMetaFile(const Path& assetFile, const AssetMeta& meta) {
+        Path metaFile = GetMetaFile(assetFile);
+        std::ofstream stream(metaFile);
+        json metaJSON = {
+            { "id", uuids::to_string(meta.id) },
+            { "display_name", meta.displayName }
+        };
+        stream << metaJSON.dump(4) << std::endl;
+    }
+
+    void AssetRegistry::RegisterCoreComponents() {
+        for (const auto& component: ClassRegistry::GetComponentClasses()) {
+            auto temp = component.construct();
+            if (!temp)
+                continue;
+
+            auto* tempComponent = temp.try_cast<IComponent>();
+            if (tempComponent == nullptr)
+                continue;
+
+            UUID assetID = UUIDUtils::Generate();
+            auto name = tempComponent->GetClassName();
+            m_assetNames_IDs[name] = assetID;
             m_assetMeta[assetID] = {
                 .id = assetID,
-                .type = assetType,
-                .name = assetName,
-                .displayName = relativeFilePath.stem().string(),
-                .path = FileSystem::absolute(file).lexically_normal().string()
+                .type = tempComponent->GetAssetType(),
+                .name = name,
+                .displayName = tempComponent->GetDisplayName(),
+                .isComponent = true
             };
-            searchPath.assets.push_back(assetID);
         }
     }
 
